[Refector] 유저 서비스 리팩토링
* 유저 관련한 엔티티, dto의 가격 관련된 필드들 타입을 int에서 BigDecimal로 변경

+++ src/main/java/kr/hhplus/be/server/application/user/dto/result/UserResult.java
@@ -3,10 +3,12 @@ package kr.hhplus.be.server.application.user.dto.result;
 
 import kr.hhplus.be.server.domain.user.dto.info.UserInfo;
 
+import java.math.BigDecimal;
+
 public record UserResult(
         Long id,
         String name,
-        int point
+        BigDecimal point
 ) {
     public static UserResult from(UserInfo userInfo) {
         return new UserResult(
