+++ src/main/java/kr/hhplus/be/server/domain/user/dto/info/UserInfo.java
@@ -2,10 +2,12 @@ package kr.hhplus.be.server.domain.user.dto.info;
 
 import kr.hhplus.be.server.domain.user.entity.User;
 
+import java.math.BigDecimal;
+
 public record UserInfo(
         Long id,
         String name,
-        int point
+        BigDecimal point
 ) {
     public static UserInfo from(User user) {
         return new UserInfo(
